Small wrapper around sqlite3
